+++ lib/features/model/cluster/config.dart
@@ -5,7 +5,7 @@
 /// License: GNU General Public License, Version 3 (the "License")
 /// https://www.gnu.org/licenses/gpl-3.0.en.html
 //
-// Time-stamp: <Sunday 2024-06-09 10:18:57 +1000 Graham Williams>
+// Time-stamp: <Wednesday 2024-06-12 12:11:05 +1000 Graham Williams>
 //
 // This program is free software: you can redistribute it and/or modify it under
 // the terms of the GNU General Public License as published by the Free Software
@@ -29,6 +29,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'package:rattle/r/source.dart';
+import 'package:rattle/widgets/activity_button.dart';
 
 /// The CLUSTER tab config currently consists of just a BUILD button.
 ///
@@ -58,12 +59,12 @@ class ClusterConfigState extends ConsumerState<ClusterConfig> {
 
             // The BUILD button.
 
-            ElevatedButton(
+            ActivityButton(
               onPressed: () {
                 rSource(ref, 'model_template');
                 rSource(ref, 'model_build_cluster');
               },
-              child: const Text('Build'),
+              child: const Text('Build Clustering'),
             ),
           ],
         ),